CalculationsService: Changed types to float

+++ calculations-service/src/main/java/com/dzydowicz/calculations/application/CalculationsServicePort.java
@@ -1,5 +1,5 @@
 package com.dzydowicz.calculations.application;
 
 public interface CalculationsServicePort {
-    Double calculateKineticEnergy(Integer mass);
+    Float calculateKineticEnergy(Float mass);
 }
